refactor: split App.jsx into components (#46)
Extract pages and map into dedicated components:
- HomePage, ResultsPage, ItineraryPage in src/components/
- ResultsMap, ItineraryMap in src/components/
- Shared constants in src/constants.js
- Shared helpers in src/utils.js
App.jsx now only manages state and event handlers (~110 lines)
